fix: resolve CI failures for Hytte-j3ll
Move appendAnswerDigit out of MathAnswerPad.tsx into mathUtils.ts so the component file only exports React components, satisfying the react-refresh/only-export-components ESLint rule.
